==> 1_Data_Cleaning.ipynb <==
import pandas as pd


# Alzheimers data

# Load the CSV file into a DataFrame
alzheimers_data = pd.read_csv('/Users/Student/Library/Mobile Documents/com~apple~CloudDocs/SETU/Sem 2/Data Analytics and Algorithms/Assignment/Implementation/Data-Science-Implementation/Data/alzheimers.csv')
alzheimers_data.head()

# Check the number of rows and columns in the DataFrame
num_rows, num_columns = alzheimers_data.shape
num_rows, num_columns

# Display basic information about the dataset
alzheimers_data.info()

# Display summary statistics of the dataset
alzheimers_data.describe()

# Check for missing values
missing_values = alzheimers_data.isnull().sum()
print("Missing values in each column:\n", missing_values)

# Check if the data is imbalanced
diagnosis_counts = alzheimers_data['Diagnosis'].value_counts()
print("Diagnosis counts:\n", diagnosis_counts)

# Determine if the data is imbalanced
if diagnosis_counts.max() / diagnosis_counts.min() > 1.5:
    print("The data is imbalanced.")
    print("Imbalanced data:\n", diagnosis_counts)
else:
    print("The data is not imbalanced.")

# Check for duplicate rows in the dataset
duplicate_rows = alzheimers_data.duplicated()
num_duplicates = duplicate_rows.sum()

if num_duplicates > 0:
    print("There are duplicates in alzheimers dataset.")
else:
    print("There are no duplicates in alzheimers dataset.")

# Save the cleaned data to a CSV file
alzheimers_data.to_csv('alzheimers_cleaned_data.csv', index=False)

# Student Performance data

# Load the Student Performance CSV file into a DataFrame
student_performance_data = pd.read_csv('/Users/Student/Library/Mobile Documents/com~apple~CloudDocs/SETU/Sem 2/Data Analytics and Algorithms/Assignment/Implementation/Data-Science-Implementation/Data/Student_Performance.csv')
student_performance_data.head()

# Check the number of rows and columns in the DataFrame
num_rows, num_columns = student_performance_data.shape
num_rows, num_columns

# Display basic information about the dataset
student_performance_data.info()

# Display summary statistics of the student performance dataset
student_performance_data.describe()

# Check for missing values in the student performance dataset
missing_values_student_performance = student_performance_data.isnull().sum()
print("Missing values in each column of student performance data:\n", missing_values_student_performance)

# Check for duplicate rows in the student performance dataset
duplicate_rows_student_performance = student_performance_data.duplicated()
num_duplicates_student_performance = duplicate_rows_student_performance.sum()

if num_duplicates_student_performance > 0:
    print("There are duplicates in the student performance dataset.")
else:
    print("There are no duplicates in the student performance dataset.")

# Save the cleaned data to a CSV file
student_performance_data.to_csv('student_performance_cleaned_data.csv', index=False)
